requirements.txt
Flask==3.1.0
Pillow==11.0.0
numpy==2.1.3
Werkzeug==3.1.3
click==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
blinker==1.9.0
